Hooks/Icecast YP: Fix a bug in addToDir()
The code that determines the hostname to use has a small issue in one
of its checks, the one which checks if the hostname contains a custom
port (for example, "http://foo:8000"). It used the array .split()
returned to check if there's an explicit port, but the if condition
was wrong.